requirements.txt
aiohttp==3.7.4.post0
nest-asyncio==1.5.1
Pillow==8.1.0
more-itertools==8.10.0
albumentations==1.0.0
torch==1.7.1
torchvision==0.8.2
transformers==4.11.3
opencv-python==4.5.3.56
numpy==1.20.2